refactor(RequestHandler): 요청을 처리하는 부분을 RequestHandler에게 위임
