+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.connector;
 
-import org.apache.catalina.Session;
 import org.apache.coyote.http11.Http11Processor;
+import org.apache.coyote.http11.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,15 +18,15 @@ public class Connector implements Runnable {
     private static final int DEFAULT_ACCEPT_COUNT = 100;
 
     private final ServerSocket serverSocket;
-    private final Session session;
+    private final RequestHandler requestHandler;
     private boolean stopped;
 
-    public Connector(final Session session) {
-        this(DEFAULT_PORT, DEFAULT_ACCEPT_COUNT, session);
+    public Connector(final RequestHandler requestHandler) {
+        this(DEFAULT_PORT, DEFAULT_ACCEPT_COUNT, requestHandler);
     }
 
-    public Connector(final int port, final int acceptCount, final Session session) {
-        this.session = session;
+    public Connector(final int port, final int acceptCount, final RequestHandler requestHandler) {
+        this.requestHandler = requestHandler;
         this.serverSocket = createServerSocket(port, acceptCount);
         this.stopped = false;
     }
@@ -69,7 +69,7 @@ public class Connector implements Runnable {
         if (connection == null) {
             return;
         }
-        var processor = new Http11Processor(connection, session);
+        var processor = new Http11Processor(connection, requestHandler);
         new Thread(processor).start();
     }
